Standardize meaning of rare
Whether two or three GET arrows are lit only changes the pool of available pokemon, not their spawn rates. See for example Slugma and Torchic, which are both common (starting weight of 10), but only appear with 2 and 3 arrows respectively. Rare should be reserved for pokemon with a starting weight of 1